autotest add test mapreduce

+++ autotest.sh
@@ -1,6 +1,14 @@
-# test shardkv
+# test mapreduce
+cd ./mapreduce/main
+bash test-mr.sh > ../../logs/mapreduce.log 2>&1
+cd ../../
+
 cd ./shardkv
-rm ../logs/shardkv.log
+# test shardkv
+if [ -f ../logs/shardkv.log ]; then
+    echo "文件存在"
+    rm ../logs/shardkv.log
+fi
 for i in {1..100}
 do
     echo "第 $i 次循环" >> ../logs/shardkv.log
